Address review comments from Jean (thanks!)
